fixed party menu display for learnable moves, added wild encounter manipulation with ability

+++ include/item.h
@@ -69,7 +69,7 @@ u8 ItemId_GetPocket(u16 itemId);
 u8 ItemId_GetType(u16 itemId);
 ItemUseFunc ItemId_GetFieldFunc(u16 itemId);
 u8 ItemId_GetBattleUsage(u16 itemId);
-u8 ItemId_GetSecondaryId(u16 itemId);
+u16 ItemId_GetSecondaryId(u16 itemId);
 u32 ItemId_GetFlingPower(u32 itemId);
 u16 ItemId_GetPrice(u16 itemId);
 void ClearBag(void);
