+++ src/item.c
@@ -688,7 +688,7 @@ bool8 ItemId_GetBattleUsage(u16 itemId)
         return gItems[item].battleUsage;
 }
 
-u8 ItemId_GetSecondaryId(u16 itemId)
+u16 ItemId_GetSecondaryId(u16 itemId)
 {
     return gItems[SanitizeItemId(itemId)].secondaryId;
 }
